refactor(cli): extract scan progress helpers into cli/scan_progress.py
